+++ core/blockchain.ts/CHANGELOG.md
@@ -10,4 +10,4 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 - feat(api): add Reader and Writer classes to handle buffering
 - feat(api): add a Template engine based on Writer class
 - feat(api): add abstract classes Block, Transaction and Wallet
-- feat(api): add MerkleTree class
+- feat(api): add MerkleTree structure class and tests